# backend/internal/docs/swagger.yaml
definitions:
...
  handler.PublishVideoRequest:
    properties:
      cover_url:
        type: string
      description:
        type: string
      play_url:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - cover_url
    - play_url
    - title
    type: object
...
  handler.UpdateVideoRequest:
    properties:
      cover_url:
        type: string
      description:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: 高性能短视频 Feed 流系统 API 文档
  termsOfService: http://swagger.io/terms/
  title: TideFlow 短视频 Feed 流 API
  version: v1
paths:
...
  /api/v1/videos:
    post:
      consumes:
      - application/json
      description: 发布一个新视频
      parameters:
      - description: 视频信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.PublishVideoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 发布视频
      tags:
      - 视频
  /api/v1/videos/{id}:
    delete:
      description: 软删除视频，仅作者可删除
      parameters:
      - description: 视频ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 删除视频
      tags:
      - 视频
    get:
      description: 通过视频ID获取视频详细信息
      parameters:
      - description: 视频ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取视频详情
      tags:
      - 视频
    put:
      consumes:
      - application/json
      description: 更新视频的标题、描述、封面或标签
      parameters:
      - description: 视频ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateVideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 更新视频信息
      tags:
      - 视频
...
  /api/v1/videos/cover:
    post:
      consumes:
      - multipart/form-data
      description: 上传视频封面图片，返回封面URL
      parameters:
      - description: 封面图片
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 上传封面
      tags:
      - 视频
  /api/v1/videos/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传视频文件，返回播放URL
      parameters:
      - description: 视频文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - OAuth2Password: []
      summary: 上传视频文件
      tags:
      - 视频
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      read: Grant read access
      write: Grant write access
    tokenUrl: /api/v1/auth/oauth2/token
    type: oauth2
swagger: "2.0"